allow inline tagging of messages
There is kind of a hack to allow doing something sensible when the first
message to start a thread is just tags, namely have the thread begin
with an empty message.  This is gross, but it is fairly hacky in other
places to allow the creation of a thread without any messages, as well
as somewhat nonsensical.  This also requires passing an option to
new-message-view so it can tell if it is a new conversation, which is
kind of gross.