subscribe now uses a multimethod, not register
It seemed like it was extra work to have to both define a function in
state and then also register the sub in another namespace.  Reframe
works that way because it's a library, so it can't rely on the user
overriding their multimethod, but for our purposes, I think this way is
easier.

This refactor also simplifies the state functions by making the called
functions always get all their arguments in one vector, instead of a
vector of constant arguments and a vector of dynamic arguments, since
that distinction only matters to `subscribe`, not to the actual state
functions.