Make searching for users uglier but working & show card on search
Previously, clicking the user name would search for `@user-name`,
which is unlikely to actually work. It now searches for `@<user-id>`,
which will actually find mentions in message text -- this still may
not be exactly what users expect (i.e. search for any messages from
that user) but it is an easy client-side change that is better than
the previous behaviour.

Also, we now show the "card" that shows up in hover of a user on the
search page when searching for a user.