Make transit writer for bots not use write caching
Having write caching enabled makes multiple occurances of keywords get
replaced with `^nn`, where `nn` is the count of "cacheable things"
that the writer saw.

This makes deserializing things quite a pain in the ass, as it
requires keeping track of the order in which keys appear and generally
complicates parsing for the bots quite a bit. Therefore, we construct
a fake WriteCache implementation that declares everything uncacheable,
then construct the gunk around that that the default writer function
would do.