Move users in app state inside groups
That is, instead of having users at the top-level of app state & have
a list of the known groups they're in, just duplicate the users under
each group. This both makes getting users in the current group much
more straightforward, but it opens the way to having per-group user profiles.