When user updates name, use broadcast-group-change helper
Previously, when using broadcast-user-change, the server would send a
list of all groups the user was in; however, this means that it would
include groups that recipients wouldn't necessarily know about. This
caused an issue when said recipients would then assoc-in to a group
they didn't know about, which made a weird phantom group show up in
the sidebar.

Closes #168