Fix thread notices and mentions not visually updating
Issue was that new implementation of threads-view (properly) wouldn't touch existing threads, so the prop thread-view would receive would not change, and hence thread-view would not re-render.

Messages would still update, because those are subscribed-to seperately.

Change was to make thread-view take just the thread-id and to subscribe to the entire thread object.