[chat] Fix opening of recent threads for users joining a group
Existing code was not working, because it was trying to open threads within the same transaction that was adding the user to the group (and as of the start of that transaction, none of the threads are visible to the user).

Fix is to run the thread opening in a seperate transaction.